Rename page folders to camelCase
