+++ src/components/pages/frontPage/FrontPage.tsx
