+++ src/components/pages/frontPage/frontPage.scss
