+++ src/components/pages/thankYouPage/ThankYouPage.scss
