+++ src/components/pages/thankYouPage/index.tsx
